consolidate __process_java_worker_classpath into build_java_worker_command
