+++ python/ray/scripts/scripts.py
@@ -10,7 +10,6 @@ import os
 import subprocess
 import sys
 import time
-import re
 
 import ray.services as services
 from ray.autoscaler.commands import (
@@ -45,40 +44,6 @@ def check_no_existing_redis_clients(node_ip_address, redis_client):
         if ray.utils.decode(info[b"node_ip_address"]) == node_ip_address:
             raise Exception("This Redis instance is already connected to "
                             "clients with this IP address.")
-
-
-def __process_java_worker_classpath(java_options):
-    """Add ray jars path to java classpath
-
-    Args:
-        java_options: java options specified by `--java-worker-options`
-    Returns:
-        java options with ray jars appended to classpath.
-    """
-    try:
-        from ray_java import resource_util
-        ray_jars_dir = resource_util.get_ray_jars_dir()
-    except ModuleNotFoundError:
-        raise Exception("Please install `ray-java` package "
-                        "when `--include-java` is enabled.")
-    cp_sep = ":"
-    import platform
-    if platform.system() == "Windows":
-        cp_sep = ";"
-    if java_options is None:
-        java_options = ""
-    options = re.split("\\s+", java_options)
-    cp_index = -1
-    for i in range(len(options)):
-        option = options[i]
-        if option == "-cp" or option == "-classpath":
-            cp_index = i + 1
-            break
-    if cp_index != -1:
-        options[cp_index] = options[cp_index] + cp_sep + ray_jars_dir
-    else:
-        options = ["-cp", ray_jars_dir] + options
-    return " ".join(options)
 
 
 @click.group()
@@ -302,9 +267,6 @@ def start(node_ip_address, redis_address, address, redis_port,
 
     redirect_worker_output = None if not no_redirect_worker_output else True
     redirect_output = None if not no_redirect_output else True
-    if include_java:
-        java_worker_options = __process_java_worker_classpath(
-            java_worker_options)
     ray_params = ray.parameter.RayParams(
         node_ip_address=node_ip_address,
         object_manager_port=object_manager_port,
