+++ python/ray/services.py
@@ -8,6 +8,7 @@ import logging
 import multiprocessing
 import os
 import random
+import re
 import resource
 import socket
 import subprocess
@@ -1252,6 +1253,40 @@ def start_raylet(redis_address,
     return process_info
 
 
+def __process_java_worker_classpath(java_options):
+    """Add ray jars path to java classpath
+
+    Args:
+        java_options: java options specified by `--java-worker-options`
+    Returns:
+        java options with ray jars appended to classpath.
+    """
+    try:
+        from ray_java import resource_util
+        ray_jars_dir = resource_util.get_ray_jars_dir()
+    except ModuleNotFoundError:
+        raise Exception("Please install `ray-java` package "
+                        "when `--include-java` is enabled.")
+    cp_sep = ":"
+    import platform
+    if platform.system() == "Windows":
+        cp_sep = ";"
+    if java_options is None:
+        java_options = ""
+    options = re.split("\\s+", java_options)
+    cp_index = -1
+    for i in range(len(options)):
+        option = options[i]
+        if option == "-cp" or option == "-classpath":
+            cp_index = i + 1
+            break
+    if cp_index != -1:
+        options[cp_index] = options[cp_index] + cp_sep + ray_jars_dir
+    else:
+        options = ["-cp", ray_jars_dir] + options
+    return " ".join(options)
+
+
 def build_java_worker_command(
         java_worker_options,
         redis_address,
@@ -1274,8 +1309,6 @@ def build_java_worker_command(
     Returns:
         The command string for starting Java worker.
     """
-    assert java_worker_options is not None
-
     command = "java "
 
     if redis_address is not None:
@@ -1298,10 +1331,34 @@ def build_java_worker_command(
     command += ("-Dray.raylet.config.num_workers_per_process_java=" +
                 "RAY_WORKER_NUM_WORKERS_PLACEHOLDER ")
 
-    if java_worker_options:
-        # Put `java_worker_options` in the last, so it can overwrite the
-        # above options.
-        command += java_worker_options + " "
+    # Add ray jars path to java classpath
+    try:
+        from ray_java import resource_util
+        ray_jars_dir = resource_util.get_ray_jars_dir()
+    except ModuleNotFoundError:
+        raise Exception("Please install `ray-java` package "
+                        "when `--include-java` is enabled.")
+    cp_sep = ":"
+    import platform
+    if platform.system() == "Windows":
+        cp_sep = ";"
+    if java_worker_options is None:
+        java_worker_options = ""
+    options = re.split("\\s+", java_worker_options)
+    cp_index = -1
+    for i in range(len(options)):
+        option = options[i]
+        if option == "-cp" or option == "-classpath":
+            cp_index = i + 1
+            break
+    if cp_index != -1:
+        options[cp_index] = options[cp_index] + cp_sep + ray_jars_dir
+    else:
+        options = ["-cp", ray_jars_dir] + options
+    java_worker_options = " ".join(options)
+    # Put `java_worker_options` in the last, so it can overwrite the
+    # above options.
+    command += java_worker_options + " "
 
     command += "RAY_WORKER_DYNAMIC_OPTION_PLACEHOLDER_0 "
     command += "org.ray.runtime.runner.worker.DefaultWorker"
